requirements.txt
aiogram>=3.4
yt-dlp
openai>=1.12
anthropic>=0.18
python-docx>=1.1
python-dotenv>=1.0
pydub>=0.25
